+++ resources/views/forms_create/empresas_destinatarias.blade.php
@@ -46,7 +46,7 @@
                     <!-- Botão de Editar -->
                     <a class="far fa-edit" href="{{route('destinatarias_edit',$destinataria->id_empresa_destinataria)}}" method="GET"></a>
                     <!-- Botão de Apagar -->
-                    <a class="fas fa-trash" href="/destinataria/delete/{{$destinataria->id_empresa_destinataria}}" method="GET"></a>
+                    <a class="fas fa-trash" href="/destinataria/delete/{{$destinataria->id_empresa_destinataria}}" onclick="return confirm('Deseja realmente excluir?')" method="GET"></a>
                 </td>
