+++ resources/views/forms_create/job.blade.php
@@ -49,7 +49,7 @@
                     <!-- Botão de Editar -->
                     <a class="far fa-edit" href="{{route('job_edit', $jobs->id_job)}}" method="GET"></a>
                     <!-- Botão de Apagar -->
-                    <a class="fas fa-trash" href="/job/delete/{{$jobs->id_job}}" method="GET"></a>
+                    <a class="fas fa-trash" href="/job/delete/{{$jobs->id_job}}" onclick="return confirm('Deseja realmente excluir?')" method="GET"></a>
                 </td>
