+++ resources/views/forms_create/origens.blade.php
@@ -48,7 +48,7 @@
                     <!-- Botão de Editar -->
                     <a class="far fa-edit" href="{{route('origem_edit', $origens->id_origem)}}" method="GET"></a>
                     <!-- Botão de Apagar -->
-                    <a class="fas fa-trash" href="/origem/delete/{{$origens->id_origem}}" method="GET"></a>
+                    <a class="fas fa-trash" href="/origem/delete/{{$origens->id_origem}}" onclick="return confirm('Deseja realmente excluir?')" method="GET"></a>
                     
                 </td>
             </tr>
