+++ resources/views/forms_search/documentos_search.blade.php
@@ -195,11 +195,11 @@
 
                     <!-- Botão de Editar -->
 
-                    <a class="far fa-edit" href="documentos_edit/{{$dashboard->id_codigo}}" method="GET"></a>
+                    <a class="far fa-edit" href="documentos_edit/{{$dashboard->id_codigo}}" method="GET" onclick=""></a>
 
                     <!-- Botão de Apagar -->
 
-                    <a class="fas fa-trash" href="delete/{{$dashboard->id_codigo}}" method="GET"></a>
+                    <a class="fas fa-trash" href="delete/{{$dashboard->id_codigo}}" onclick="return confirm('Deseja realmente excluir?')" method="GET"></a>
 
                     <!-- Botão de Anexo -->
                     <form method="POST" action="{{route('visualizar_anexo')}}" >
